requirements.txt
absl-py==1.0.0
astor==0.8.1
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.6.20
chardet==5.0.0
click==8.0.4
commonmark==0.9.1
configparser==5.2.0
contextlib2==21.6.0
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
gast==0.5.3
google-pasta==0.2.0
grad-cam==1.4.8
grpcio==1.43.0
h5py==3.1.0
imageio==2.13.5
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipdb==0.13.13
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.1.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.9.3
Markdown==3.3.6
matplotlib==3.3.4
networkx==2.5.1
nltk==3.6.7
numpy==1.19.5
nvidia-ml-py==11.525.131
nvitop==1.0.0
opencv-python==4.5.5.62
packaging==21.3
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.15.0
prompt-toolkit==3.0.36
protobuf==3.19.3
psutil==5.9.5
ptyprocess==0.7.0
Pygments==2.14.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.1.1
PyYAML==6.0
regex==2023.8.8
rich==12.6.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
Shapely==1.8.5.post1
SimpleITK==2.0.2
six==1.16.0
summary==0.2.0
tenacity==8.2.2
tensorboard==1.14.0
tensorboardX==2.6.2
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
tomli==1.2.3
torch==1.9.1+cu111
torchaudio==0.9.1
torchcam==0.3.2
torchsummary==1.5.1
torchvision==0.10.1+cu111
tqdm==4.64.1
traitlets==4.3.3
ttach==0.0.3
typing_extensions==4.0.1
wcwidth==0.2.6
Werkzeug==2.0.2
wrapt==1.13.3
xgboost==1.5.2
zipp==3.6.0
